Moved all C++ functions into separate files with corresponding header files + Don't use RcppAttributes anymore for importing RcppArmadillo, etc.
