+++ src/est_betak.h
@@ -1,0 +1,8 @@
+#ifndef EST_BETAK_H
+#define EST_BETAK_H
+
+#include <RcppArmadillo.h>
+
+arma::rowvec est_betak(const arma::rowvec& wz_co, float gamma_);
+
+#endif
