--- src/functions.cpp
@@ -1,1426 +1,0 @@
-#include <RcppArmadilloExtensions/sample.h>
-#include <iostream>
-#include <vector>
-#include <progress.hpp>
-#include <progress_bar.hpp>
-
-#include "error.h"
-#include "rmvnorm_cpp.h"
-#include "table_cpp.h"
-#include "draw_eta_norm.h"
-#include "invlogit.h"
-#include "get_loglike.h"
-
-using namespace Rcpp;
-// [[Rcpp::depends(RcppArmadillo)]]
-// [[Rcpp::depends(RcppProgress)]]
-// [[Rcpp::plugins(cpp11)]]
-
-//' Log-posterior for regression coefficients with normal prior
-//'
-//' @param ll A double of the current log-likelihood.
-//' @param eta A q x 1 vector of regression coefficients.
-//' @param mu0 A q x 1 mean vector for the prior on the regression coefficients.
-//' @param sigma0 A q x q variance-covariance matrix for the prior on the
-//'   regression coefficients.
-//'
-//' @return The current log-posterior.
-double get_lpost_eta(double ll, const arma::colvec& eta,
-                     const arma::colvec& mu0, const arma::mat& sigma0) {
-
-  double lp_temp = ll;
-  // Add prior on eta
-  double temp_prod = arma::as_scalar(
-    (eta - mu0).t() * arma::inv_sympd(sigma0) * (eta - mu0)
-  );
-  lp_temp += (-0.5 * temp_prod);
-
-  return lp_temp;
-}
-
-//' Compute full conditional log-posterior of eta for logistic sLDA/sLDAX/regression
-//'
-//' @param w A D x q matrix containing a predictor model matrix of assumed form
-//'   (X, Zbar, XZbarInteractions).
-//' @param y A D x 1 vector of the outcome variable for each document.
-//' @param eta A q x 1 vector of regression coefficients
-//' @param mu0 A q x 1 vector of prior means for the regression coefficients.
-//' @param sigma0 A q x q prior variance-covariance matrix for the regression
-//'   coefficients.
-//'
-//' @return The full conditional log-posterior density of eta.
-double eta_logpost_logit(const arma::mat& w, const arma::colvec& y,
-                         const arma::vec& eta,
-                         const arma::vec& mu0, const arma::mat& sigma0) {
-
-  // Compute log-likelihood of y
-  double loglike = get_ll_logit(y, w, eta);
-  // Add log-prior on eta
-  double logpost = get_lpost_eta(loglike, eta, mu0, sigma0);
-  return logpost;
-}
-
-//' Draw eta from full conditional posterior for logistic sLDA/sLDAX/regression
-//'   using Metropolis-Hastings (MH) algorithm
-//'
-//' @param w A D x q matrix containing a predictor model matrix of assumed form
-//'   (X, Zbar, XZbarInteractions).
-//' @param y A D x 1 vector of the outcome variable for each document.
-//' @param eta_prev A q x 1 vector of the previous draw of the regression
-//'   coefficients.
-//' @param mu0 A q x 1 vector of prior means for the regression coefficients.
-//' @param sigma0 A q x q prior variance-covariance matrix for the regression
-//'   coefficients.
-//' @param proposal_sd A q x 1 vector of proposal distribution standard deviations.
-//' @param attempt A vector of the number of current attempted draws of eta.
-//' @param accept A vector of the number of accepted draws of eta.
-//'
-//' @return A q x 1 vector of draws for eta.
-arma::colvec draw_eta_logit(const arma::mat& w, const arma::colvec& y,
-                         const arma::colvec& eta_prev,
-                         const arma::colvec& mu0, const arma::mat& sigma0,
-                         const arma::vec& proposal_sd,
-                         arma::vec& attempt, arma::vec& accept) {
-
-  const uint16_t q = w.n_cols;
-  arma::colvec cand_eta = eta_prev; // Candidate draws of eta
-  arma::colvec eta = eta_prev;
-  double cur_logpost = eta_logpost_logit(w, y, eta_prev, mu0, sigma0);
-
-  for (uint16_t j = 0; j < q; j++) {
-    cand_eta(j) = rnorm(1, eta_prev(j), proposal_sd(j))(0);
-    attempt(j)++; // Passed by reference to update outside function
-
-    // Compute acceptance ratio
-    double cand_logpost = eta_logpost_logit(w, y, cand_eta, mu0, sigma0);
-    double log_r = cand_logpost - cur_logpost; // Symmetric proposals
-    double log_u = log(runif(1)(0));
-    if (log_r > log_u) {
-      eta(j)      = cand_eta(j);
-      cur_logpost = cand_logpost;
-      accept(j)++; // Passed by reference to update outside function
-    } else {
-      eta(j)      = eta_prev(j);
-      cand_eta(j) = eta_prev(j);
-    }
-  }
-  return eta;
-}
-
-//' Draw sigma2 from full conditional posterior for sLDA/sLDAX/MLR
-//'
-//' @param a0 The prior shape parameter for \eqn{\sigma^2}.
-//' @param b0 The prior scale parameter for \eqn{\sigma^2}.
-//' @param w A D x q matrix containing a predictor model matrix of assumed form
-//'   (X, Zbar, XZbarInteractions).
-//' @param y A D x 1 vector of the outcome variable.
-//' @param eta A q x 1 vector of regression coefficients.
-//'
-//' @return A draw for \eqn{\sigma^2}.
-long double draw_sigma2(float a0, float b0,
-                        const arma::mat& w, const arma::colvec& y,
-                        const arma::colvec& eta) {
-
-  const uint32_t D = y.size(); // Number of observations
-  if ((a0 <= 0.0) || (b0 <= 0.0)) stop("a0 and b0 must be positive");
-  long double a = 0.5 * (static_cast<float>(D) + a0);
-
-  if ((w.n_rows != D) || (w.n_cols != eta.size()))
-    error("zbar must be a D x q matrix and eta must be a q x 1 vector");
-
-  arma::colvec resid(D);
-  resid = y - w * eta;
-  double b_update = arma::as_scalar(resid.t() * resid);
-
-  // Parameterization is 1 / rate
-  long double b = 1.0 / (0.5 * (b0 + b_update));
-  long double sigma2inv = R::rgamma(a, b);
-  return 1.0 / sigma2inv;
-}
-
-// Estimate \eqn{\beta_k}
-//
-// @param wz_co A V x 1 vector of counts of the draws of each word for topic
-//   k over all documents.
-// @param gamma_ The hyperparameter for the Dirichlet priors on \eqn{\beta_k}.
-//
-// @return A V x 1 vector of estimates for \eqn{\beta_k}.
-//' @export
-// [[Rcpp::export(.est_betak)]]
-arma::rowvec est_betak(const arma::rowvec& wz_co, float gamma_) {
-
-  // Warning: Overflow caused by probabilities near 0 handled by setting
-  //   probability for the problematic word to 0.0;
-  if (gamma_ < 0.0) error("gamma_ must be positive");
-  const uint32_t V = wz_co.size(); // Vocabulary size
-  if (V < 2) error("vocabulary size V must be at least 2");
-
-  arma::rowvec betak = exp(log(wz_co + gamma_) - log(sum(wz_co) + V * gamma_));
-
-  // Check for impossible estimates and replace with 0
-  betak.transform( [](double val) {
-                       if ((val > 1.0) | (std::isnan(val))) {
-                         return(0.0);
-                       } else {
-                         return val;
-                       }
-  });
-
-  return betak;
-}
-
-// Estimate \eqn{\theta_d}
-//
-// @param z_count A K x 1 vector of counts of topic draw in document d.
-// @param alpha_ The hyperparameter on the Dirichlet prior for \eqn{\theta_d}.
-//
-// @return A K x 1 vector of estimate for \eqn{\theta_d}.
-//' @export
-// [[Rcpp::export(.est_thetad)]]
-arma::rowvec est_thetad(const arma::rowvec& z_count, float alpha_) {
-
-  // Warning: Overflow caused by probabilities near 0 handled by setting
-  //   probability for the problematic topic to 0.0;
-  const uint16_t K = z_count.size(); // Number of topics
-  if (alpha_ < 0.0) error("alpha_ must be positive");
-  if (K < 2) error("number of topics must be at least 2");
-
-  arma::rowvec thetad = exp(log(z_count + alpha_) - log(sum(z_count) +
-    static_cast<float>(K) * alpha_));
-
-  // Check for impossible estimates and replace with 0
-  thetad.transform( [](double val) {
-                        if ((val > 1.0) | (std::isnan(val))) {
-                          return(0.0);
-                        } else {
-                          return val;
-                        }
-  });
-
-  return thetad;
-}
-
-// Count topic-word co-occurences in corpus.
-//
-// Computes topic-word co-occurence matrix for a corpus of \eqn{D} documents
-// with the maximum length of a document in the corpus equal to max(\eqn{N_d})
-// and a vocabulary of \eqn{V} unique terms in the corpus.
-//
-// Indices in \code{doc_topic} and \code{doc_word} where no word exists in the
-// document must be set to 0.
-//
-// @param K The number of topics.
-// @param V The number of terms in the corpus vocabulary.
-// @param doc_topic A \eqn{D} x max(\eqn{N_d}) matrix of topic assignments for
-//   the corpus.
-// @param doc_word A \eqn{D} x max(\eqn{N_d}) matrix of words for corpus.
-//
-// @return A \eqn{K} x \eqn{V} matrix of topic-word co-occurence counts.
-//' @export
-// [[Rcpp::export(.count_topic_word)]]
-arma::mat count_topic_word(uint16_t K, uint32_t V,
-                           const arma::umat& doc_topic,
-                           const arma::umat& doc_word) {
-
-  const uint32_t D = doc_topic.n_rows; // Number of documents
-  if (K < 2) error("number of topics must be at least 2");
-  if (V < 2) error("size of vocabulary V must be at least 2");
-  const uint32_t maxnd = doc_word.n_cols; // Maximum document length
-  if (doc_word.n_rows != D) error("'doc_word' and 'doc_topic' must have the same number of rows");
-  if (doc_topic.n_cols != maxnd)
-    error("'doc_topic' and 'doc_word' must have the same number of columns");
-
-  const arma::ucolvec topics_index = arma::linspace<arma::ucolvec>(1, K, K);
-  const arma::ucolvec docs_index = arma::linspace<arma::ucolvec>(0, D - 1, D);
-  // Matrix to store number of topic/word co-occurences
-  arma::mat topic_word_freq = arma::mat(K, V, arma::fill::zeros);
-
-  for (uint16_t topic : topics_index) {
-    arma::uvec ids = arma::find(doc_topic == topic);
-    arma::urowvec word_ids = doc_word.elem(ids).t();
-    // Update word frequencies for topic
-    std::map<uint32_t, uint32_t> counts = table_cpp(word_ids);
-    for (const auto &p : counts) {
-      topic_word_freq(topic - 1, p.first - 1) += p.second;
-    }
-  } // Topics
-  return topic_word_freq;
-}
-
-//' Compute log-numerator vector for sampling zdn from full conditional distribution for LDA
-//'
-//' @param V The number of terms in the corpus vocabulary.
-//' @param ndk_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} in
-//'   document d excluding the current word \eqn{w_n} from the counts.
-//' @param nkm_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} and
-//'   word m in the corpus excluding the current word \eqn{w_n} from the
-//'   counts.
-//' @param nk_n A K x 1 vector of counts of draws of topic
-//'   \eqn{k = 1, \ldots, K} in the corpus excluding the current word \eqn{w_n}
-//'   from the counts.
-//' @param alpha_ The hyperparameter on the Dirichlet prior for \eqn{\theta_d}.
-//' @param gamma_ The hyperparameter for the Dirichlet priors on \eqn{\beta_k}.
-//'
-//' @return A K x 1 vector of the log-numerator from the LDA model to sample zdn.
-arma::vec get_log_numer_samplez(uint32_t V, const arma::vec& ndk_n,
-                                const arma::vec& nkm_n, const arma::vec& nk_n,
-                                float alpha_, float gamma_) {
-
-  const uint16_t K = nk_n.size(); // Number of topics
-  if (K < 2) error("number of topics must be at least 2");
-  if (V < 2) error("size of vocabulary V must be at least 2");
-  if (ndk_n.size() != K) error("ndk_n must be a vector of length K");
-  if (nkm_n.size() != K) error("nkm_n must be a vector of length K");
-  if (nk_n.size() != K) error("nk_n must be a vector of length K");
-  if (alpha_ < 0.0) error("alpha_ must be positive");
-  if (gamma_ < 0.0) error("gamma_ must be positive");
-
-  arma::vec log_num =
-    log(ndk_n + alpha_) +
-    log(nkm_n + gamma_) -
-    log(nk_n + static_cast<float>(V) * gamma_);
-
-  return log_num;
-}
-
-//' Draw zdn from full conditional distribution for LDA/sLDA/sLDAX
-//'
-//' @param log_num A K x 1 vector of the log-numerator for sampling from topics
-//'   1, ..., K.
-//'
-//' @return Indicator for the topic draw from {1, 2, ..., K}.
-uint16_t draw_zdn(arma::vec& log_num) {
-
-  const uint16_t K = log_num.size(); // Number of topics
-  long double denom = sum(exp(log_num));
-  arma::vec pmf = exp(log_num - log(denom));
-
-  bool good_pmf = true;
-  for (uint16_t k = 0; k < K; k++) {
-    if (std::isnan(pmf(k)) || std::isinf(pmf(k))) good_pmf = false;
-  }
-
-  if (!good_pmf) pmf = 1.0 / static_cast<float>(K);
-  IntegerVector topics = seq_len(K);
-  IntegerVector zdn = RcppArmadillo::sample(topics, 1, true, pmf);
-  return zdn(0);
-}
-
-//' Draw zdn from full conditional distribution for LDA
-//'
-//' @param V The number of terms in the corpus vocabulary.
-//' @param ndk_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} in
-//'   document d excluding the current word \eqn{w_n} from the counts.
-//' @param nkm_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} and
-//'   word m in the corpus excluding the current word \eqn{w_n} from the
-//'   counts.
-//' @param nk_n A K x 1 vector of counts of draws of topic
-//'   \eqn{k = 1, \ldots, K} in the corpus excluding the current word \eqn{w_n}
-//'   from the counts.
-//'
-//' @return Indicator for the topic draw from {1, 2, ..., K}.
-uint16_t draw_zdn_lda(uint32_t V,
-                      const arma::vec& ndk_n, const arma::vec& nkm_n,
-                      const arma::vec& nk_n, float alpha_, float gamma_) {
-
-  // Warning: Overflow caused by probabilities near 0 handled by setting
-  //   probability for the problematic topic to 1 / K
-  arma::vec log_num = get_log_numer_samplez(V, ndk_n, nkm_n, nk_n,
-                                            alpha_, gamma_);
-  uint16_t zdn = draw_zdn(log_num);
-  return zdn;
-}
-
-//' Draw zdn from full conditional distribution for sLDA/sLDAX
-//'
-//' @param yd The outcome variable for document \eqn{d}.
-//' @param w_d A q x 1 vector containing row d of the predictor model matrix of
-//'   assumed form (X, Zbar, XZbarInteractions).
-//' @param eta A q x 1 vector of regression coefficients.
-//' @param sigma2 The residual variance.
-//' @param V The number of terms in the corpus vocabulary.
-//' @param ndk_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} in
-//'   document d excluding the current word \eqn{w_n} from the counts.
-//' @param nkm_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} and
-//'   word m in the corpus excluding the current word \eqn{w_n} from the
-//'   counts.
-//' @param nk_n A K x 1 vector of counts of draws of topic
-//'   \eqn{k = 1, \ldots, K} in the corpus excluding the current word \eqn{w_n}
-//'   from the counts.
-//'
-//' @return Indicator for the topic draw from {1, 2, ..., K}.
-uint16_t draw_zdn_slda_norm(double yd, const arma::rowvec& w_d,
-                            const arma::vec& eta, double sigma2,
-                            uint32_t V, const arma::vec& ndk_n,
-                            const arma::vec& nkm_n, const arma::vec& nk_n,
-                            float alpha_, float gamma_) {
-
-  // Warning: Overflow caused by probabilities near 0 handled by setting
-  //   probability for the problematic topic to 1 / K;
-  //   this tends to occur if sigma^2 draw near 0
-  if (sigma2 < 0.0) error("sigma2 must be positive");
-  arma::vec log_num = get_log_numer_samplez(V, ndk_n, nkm_n, nk_n,
-                                            alpha_, gamma_) +
-    R::dnorm(yd, dot(w_d, eta), sqrt(sigma2), true);
-  uint16_t zdn = draw_zdn(log_num);
-  return zdn;
-}
-
-//' Draw zdn from full conditional distribution for sLDA/sLDAX with binary outcome
-//'
-//' @param yd A the outcome variable for document d.
-//' @param w_d A q x 1 vector containing row d of the predictor model matrix of
-//'   assumed form (X, Zbar, XZbarInteractions).
-//' @param eta A q x 1 vector of regression coefficients.
-//' @param V The number of terms in the corpus vocabulary.
-//' @param ndk_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} in
-//'   document d excluding the current word \eqn{w_n} from the counts.
-//' @param nkm_n A K x 1 vector of counts of topic \eqn{k = 1, \ldots, K} and
-//'   word m in the corpus excluding the current word \eqn{w_n} from the
-//'   counts.
-//' @param nk_n A K x 1 vector of counts of draws of topic
-//'   \eqn{k = 1, \ldots, K} in the corpus excluding the current word \eqn{w_n}
-//'   from the counts.
-//'
-//' @return Indicator for the topic draw from {1, 2, ..., K}.
-uint16_t draw_zdn_slda_logit(double yd, const arma::rowvec& w_d,
-                             const arma::vec& eta,
-                             uint32_t V, const arma::vec& ndk_n,
-                             const arma::vec& nkm_n, const arma::vec& nk_n,
-                             float alpha_, float gamma_) {
-
-  // Warning: Overflow caused by probabilities near 0 handled by setting
-  //   probability for the problematic topic to 1 / K;
-  double muhat = dot(w_d, eta);
-  // Compute log-likelihood of y_d
-  double loglike = get_ll_logit_yd(yd, muhat);
-
-  arma::vec log_num = get_log_numer_samplez(V, ndk_n, nkm_n, nk_n,
-                                            alpha_, gamma_) +
-    loglike;
-  uint16_t zdn = draw_zdn(log_num);
-  return zdn;
-}
-
-//' Posterior predictive likelihood for sLDA/sLDAX/MLR
-//'
-//' @param w A D x q matrix of additional predictors.
-//' @param eta A q x 1 vector of regression coefficients.
-//' @param sigma2 The residual variance.
-//'
-//' @return Posterior predictive likelihood.
-arma::rowvec post_pred_norm(const arma::mat& w,
-                            const arma::colvec& eta, double sigma2) {
-
-  const uint16_t D = w.n_rows;
-  arma::colvec mu_hat = w * eta;
-
-  arma::mat sigma(D, D, arma::fill::eye); // D x D identity matrix
-  // Draw D new observations
-  arma::colvec yhat = rmvnorm_cpp(1, mu_hat, sigma2 * sigma).t();
-  arma::colvec loglike_pred = -0.5 / sigma2 * arma::square(yhat - mu_hat);
-
-  return exp(loglike_pred.t());
-}
-
-//' Posterior predictive likelihood for logistic sLDA/sLDAX/regression
-//'
-//' @param w A D x q matrix of additional predictors.
-//' @param eta A q x 1 vector of regression coefficients.
-//' @return Predictive posterior likelihood of all D observations.
-arma::rowvec post_pred_logit(const arma::mat& w, const arma::colvec& eta) {
-
-  const uint16_t D = w.n_rows;
-  arma::rowvec loglike_pred(D);
-  arma::colvec mu_hat = w * eta;
-  for (uint16_t d = 0; d < D; d++) {
-    double phat = invlogit(arma::as_scalar(mu_hat(d)));
-    uint16_t yhat = Rcpp::rbinom(1, 1, phat)(0);
-    loglike_pred(d) = get_ll_logit_yd(yhat, arma::as_scalar(mu_hat(d)));
-  }
-  return exp(loglike_pred);
-}
-
-//' Contribution to effective number of parameters for WAIC from observation y_d
-//'
-//' @param like_pred A m x 1 vector of predictive likelihoods (NOT log-likelihoods).
-//' @export
-//' @return The contribution of y_d (its predictive posterior likelihood variance)
-//'   to the effective number of parameters.
-double pwaic_d(const arma::colvec& like_pred) {
-
-  // Get variance of log-predictive likelihood
-  double mcmc_vary = arma::var(log(like_pred)); // Denominator of m - 1
-  return mcmc_vary;
-}
-
-//' WAIC for observation y_d
-//'
-//' @param like_pred A m x 1 vector of predictive likelihoods (NOT log-likelihoods) for y_d.
-//' @param p_effd The contribution to the effective number of parameters from
-//'   obs y_d.
-//' @return WAIC contribution for observation d (on deviance scale).
-double waic_d(const arma::colvec& like_pred, double p_effd) {
-
-  // Log of mean posterior predictive density for y_d
-  double lppd = log(arma::mean(like_pred));
-
-  // Contribution to WAIC (on deviance scale, i.e., -2ll) for y_d
-  return (-2.0 * (lppd - p_effd)); // See Gelman, Hwang, Vehtari (2014, p. 1003)
-}
-
-//' Compute WAIC for all outcomes.
-//'
-//' @param iter The current iteration of the chain.
-//' @param l_pred A m x D matrix of predictive likelihoods (NOT log-likelihoods).
-//'
-//' @return Vector of (1) WAIC for model, (2) standard error for WAIC, and (3)
-//'   the effective number of parameters.
-//' @export
-// [[Rcpp::export]]
-NumericVector waic_all(uint32_t iter, const arma::mat& l_pred) {
-
-  NumericVector full_waic_se(3);
-  const uint16_t D = l_pred.n_cols; // Number of observations
-  arma::colvec waic(D);
-  arma::colvec peff(D);
-
-  // Compute WAIC and p_eff for entire data set
-  for (uint16_t d = 0; d < D; d++) {
-    peff(d) = pwaic_d(l_pred.submat(0, d, iter - 1, d));
-    waic(d) = waic_d(l_pred.submat(0, d, iter - 1, d), peff(d));
-  }
-  double peff_sum = arma::sum(peff);  // p_eff
-  double waic_sum = arma::sum(waic);  // WAIC
-
-  // Compute SE(WAIC)
-  double se_waic = arma::var(waic);   // Denominator of D - 1
-  se_waic *= static_cast<float>(D);   // Variance of WAIC times D
-  se_waic = sqrt(se_waic);            // SE(WAIC)
-
-  full_waic_se(0) = waic_sum;
-  full_waic_se(1) = se_waic;
-  full_waic_se(2) = peff_sum;
-
-  return full_waic_se;
-}
-
-//' Compute difference (WAIC1 - WAIC2) in WAIC and its SE for two models.
-//'
-//' @param l_pred1 A m1 x D matrix of predictive likelihoods (NOT log-likelihoods) from model 1.
-//' @param l_pred2 A m2 x D matrix of predictive likelihoods (NOT log-likelihoods) from model 2.
-//'
-//' @return A vector of (1) the difference in WAIC (on the deviance scale)
-//'   between models and (2) the standard error of the difference in WAIC.
-//' @export
-// [[Rcpp::export]]
-NumericVector waic_diff(const arma::mat& l_pred1, const arma::mat& l_pred2) {
-
-  NumericVector diff_waic_se(2);
-  const uint32_t m1 = l_pred1.n_rows; // Length of first chain
-  const uint32_t m2 = l_pred2.n_rows; // Length of second chain
-  const uint16_t D = l_pred1.n_cols;  // Number of observations
-  arma::colvec waic1(D);
-  arma::colvec waic2(D);
-  arma::colvec peff1(D);
-  arma::colvec peff2(D);
-  arma::colvec waic_diff(D);
-
-  // Compute WAIC and p_eff for entire data set
-  for (uint16_t d = 0; d < D; d++) {
-    peff1(d) = pwaic_d(l_pred1.submat(0, d, m1 - 1, d));
-    peff2(d) = pwaic_d(l_pred2.submat(0, d, m2 - 1, d));
-    waic1(d) = waic_d(l_pred1.submat(0, d, m1 - 1, d), peff1(d));
-    waic2(d) = waic_d(l_pred2.submat(0, d, m2 - 1, d), peff2(d));
-  }
-  waic_diff = waic1 - waic2;
-  double waic_diff_sum = arma::sum(waic_diff); // WAIC1 - WAIC2
-
-  // Compute SE(WAIC1 - WAIC2)
-  double se_waic_diff = arma::var(waic_diff);  // Denominator of D - 1
-  se_waic_diff *= static_cast<float>(D);       // Multiply variance of diff by D
-  se_waic_diff = sqrt(se_waic_diff);           // SE(WAIC1 - WAIC2)
-
-  diff_waic_se(0) = waic_diff_sum; // Difference
-  diff_waic_se(1) = se_waic_diff; // SE(WAIC1 - WAIC2)
-
-  return diff_waic_se;
-}
-
-//' Log-posterior for normal outcome regression
-//'
-//' @param ll A double of the current log-likelihood.
-//' @param eta A q x 1 vector of regression coefficients.
-//' @param sigma2 The current draw of the residual variance of y.
-//' @param mu0 A q x 1 mean vector for the prior on the regression coefficients.
-//' @param sigma0 A q x q variance-covariance matrix for the prior on the
-//'   regression coefficients.
-//' @param a0 The shape parameter for the prior on sigma2.
-//' @param b0 The scale parameter for the prior on sigma2.
-//'
-//' @return The current log-posterior.
-double get_lpost_mlr(double ll,
-                     const arma::colvec& eta, double sigma2,
-                     const arma::colvec& mu0, const arma::mat& sigma0,
-                     double a0, double b0) {
-
-  double lp_temp = get_lpost_eta(ll, eta, mu0, sigma0);
-
-  // Add prior on sigma2
-  lp_temp += ((-0.5 * a0 - 1.0) * log(sigma2) - 0.5 * b0 / sigma2);
-
-  return lp_temp;
-}
-
-//' Log-posterior for LDA model
-//'
-//' @param ll A double of the current log-likelihood.
-//' @param theta A D x K matrix of the current estimates of the document topic proportions.
-//' @param beta a K x V matrix of the current estimates of the word-topic probabilities.
-//' @param gamma_ The hyper-parameter for the prior on the topic-specific
-//'   vocabulary probabilities.
-//' @param alpha_ The hyper-parameter for the prior on the topic proportions.
-//' @param V The number of words in the vocabulary.
-//' @param docs_index A vector of length D containing elements 1, 2, ..., D.
-//'
-//' @return The current log-posterior.
-double get_lpost_lda(double ll, const arma::mat& theta, const arma::mat& beta,
-                     double gamma_, double alpha_,
-                     uint32_t V, const IntegerVector& docs_index) {
-
-  // Add prior on beta matrix
-  double temp_betapost = arma::accu(log(beta));
-  double lp_temp = ll + ((gamma_ - 1.0) * temp_betapost);
-
-  // Add prior on theta matrix
-  double temp_thetapost = arma::accu(log(theta));
-  lp_temp += ((alpha_ - 1) * temp_thetapost);
-
-  return lp_temp;
-}
-
-//' Log-posterior for sLDA/sLDAX model
-//'
-//' @param ll A double of the current log-likelihood.
-//' @param eta A q x 1 vector of regression coefficients.
-//' @param sigma2 The current draw of the residual variance of y.
-//' @param theta A D x K matrix of the current estimates of the document topic proportions.
-//' @param beta a K x V matrix of the current estimates of the word-topic probabilities.
-//' @param mu0 A q x 1 mean vector for the prior on the regression coefficients.
-//' @param sigma0 A q x q variance-covariance matrix for the prior on the
-//'   regression coefficients.
-//' @param gamma_ The hyper-parameter for the prior on the topic-specific
-//'   vocabulary probabilities.
-//' @param alpha_ The hyper-parameter for the prior on the topic proportions.
-//' @param a0 The shape parameter for the prior on sigma2.
-//' @param b0 The scale parameter for the prior on sigma2.
-//' @param V The number of words in the vocabulary.
-//' @param docs_index A vector of length D containing elements 1, 2, ..., D.
-//'
-//' @return The current log-posterior.
-double get_lpost_slda_norm(double ll, const arma::colvec& eta, double sigma2,
-                           const arma::mat& theta, const arma::mat& beta,
-                           const arma::colvec& mu0, const arma::mat& sigma0,
-                           double gamma_, double alpha_,
-                           double a0, double b0, uint32_t V,
-                           const IntegerVector& docs_index) {
-
-  double lp_temp = get_lpost_mlr(ll, eta, sigma2, mu0, sigma0, a0, b0);
-  lp_temp += get_lpost_lda(lp_temp, theta, beta, gamma_, alpha_, V, docs_index);
-  return lp_temp;
-}
-
-//' Log-posterior for logistic sLDA/sLDAX model
-//'
-//' @param ll A double of the current log-likelihood.
-//' @param eta A q x 1 vector of regression coefficients.
-//' @param theta A D x K matrix of the current estimates of the document topic proportions.
-//' @param beta a K x V matrix of the current estimates of the word-topic probabilities.
-//' @param mu0 A q x 1 mean vector for the prior on the regression coefficients.
-//' @param sigma0 A q x q variance-covariance matrix for the prior on the
-//'   regression coefficients.
-//' @param gamma_ The hyper-parameter for the prior on the topic-specific
-//'   vocabulary probabilities.
-//' @param alpha_ The hyper-parameter for the prior on the topic proportions.
-//' @param V The number of words in the vocabulary.
-//' @param docs_index A vector of length D containing elements 1, 2, ..., D.
-//'
-//' @return The current log-posterior.
-double get_lpost_slda_logit(double ll, const arma::colvec& eta,
-                            const arma::mat& theta, const arma::mat& beta,
-                            const arma::colvec& mu0, const arma::mat& sigma0,
-                            double gamma_, double alpha_,
-                            uint32_t V, const IntegerVector& docs_index) {
-
-  double lp_temp = get_lpost_eta(ll, eta, mu0, sigma0);
-  lp_temp += get_lpost_lda(lp_temp, theta, beta, gamma_, alpha_, V, docs_index);
-  return lp_temp;
-}
-
-//' Update number of times topic was drawn in document excluding current word
-//'
-//' @param topic The current topic index.
-//' @param ndk A vector of the current number of draws of each topic in document d.
-//'
-//' @return A vector of the current number of draws of each topic in document d
-//'   excluding word n.
-void count_topicd(uint16_t topic, uint16_t doc, arma::mat& ndk) {
-  // Exclude word n from topic counts in doc d
-  ndk(doc, topic - 1)--;
-  if (ndk(doc, topic - 1) < 0) ndk(doc, topic - 1) = 0;
-}
-
-//' Update number of times topic was drawn in corpus excluding current word
-//'
-//' @param topic The current topic index.
-//' @param nk A vector of the current number of draws of each topic in the corpus.
-//'
-//' @return A vector of the current number of draws of each topic in the corpus
-//'   excluding the current word.
-void count_topic_corpus(uint16_t topic, arma::vec& nk) {
-  // Exclude word n from topic counts in corpus
-  nk(topic - 1)--;
-  if (nk(topic - 1) < 0) nk(topic - 1) = 0;
-}
-
-//' Update number of times word and topic co-occur in corpus
-//'
-//' @param word The current word index.
-//' @param topic The current topic index.
-//' @param nkm A K x V matrix of the current number of co-occurences of each topic
-//'   and vocabulary term in the corpus.
-//'
-//' @return A K x V matrix of the current number of co-occurences of each topic
-//'   and vocabulary term in the corpus excluding the current word.
-void count_word_topic(uint32_t word, uint16_t topic, arma::mat& nkm) {
-  // Exclude word n from topic-word counts
-  nkm(topic - 1, word - 1)--;
-  // Fix possible negative counts
-  if (nkm(topic - 1, word - 1) < 0) nkm(topic - 1, word - 1) = 0;
-}
-
-//' Update all topic and topic-word counts in document and corpus
-//'
-//' @param d The current document index.
-//' @param word The current word index.
-//' @param topic The current topic index.
-//' @param ndk A vector of the current number of draws of each topic in document d.
-//' @param nk A vector of the current number of draws of each topic in the corpus.
-//' @param nkm A K x V matrix of the current number of co-occurences of each topic
-//'   and vocabulary term in the corpus.
-void update_zcounts(uint32_t d, uint32_t word, uint16_t topic, uint32_t doc,
-                    arma::mat& ndk, arma::vec& nk, arma::mat& nkm) {
-
-  count_topicd(topic, doc, ndk);
-  count_topic_corpus(topic, nk);
-  count_word_topic(word, topic, nkm);
-}
-
-//////////////////////////////// Gibbs Samplers ////////////////////////////////
-
-// Collapsed Gibbs sampler for multiple linear regression
-//
-// @param m The number of iterations to run the Gibbs sampler.
-// @param burn The number of iterations to discard as the burn-in period.
-// @param thin The period of iterations to keep after the burn-in period
-//   (default: 1).
-// @param y A D x 1 vector of outcomes to be predicted.
-// @param x A D x (p + 1) matrix of additional predictors.
-// @param mu0 A (p + 1) x 1 mean vector for the prior on the regression
-//   coefficients.
-// @param sigma0 A (p + 1) x (p + 1) variance-covariance matrix for the
-//   prior on the regression coefficients.
-// @param eta_start A (p + 1) x 1 vector of starting values for the
-//   regression coefficients.
-// @param a0 The shape parameter for the prior on sigma2 (default: 0.001)
-// @param b0 The scale parameter for the prior on sigma2 (default: 0.001)
-// @param verbose Should parameter draws be output during sampling? (default:
-//   \code{FALSE}).
-// @param display_progress Should percent progress of sampler be displayed
-//   (default: \code{FALSE}). Recommended that only one of \code{verbose} and
-//   \code{display_progress} be set to \code{TRUE} at any given time.
-//
-// @return An object of class \code{Mlr}.
-//' @export
-// [[Rcpp::export(.gibbs_mlr_cpp)]]
-S4 gibbs_mlr_cpp(uint32_t m, uint32_t burn, uint32_t thin,
-                 const arma::colvec& y, const arma::mat& x,
-                 const arma::colvec& mu0, const arma::mat& sigma0,
-                 arma::colvec eta_start, float a0 = 0.001, float b0 = 0.001,
-                 bool verbose = false, bool display_progress = false) {
-
-  if (m <= burn) stop("Length of chain m not greater than burn-in period.");
-  if ( (thin == 0) || (thin > (m - burn)) )
-    stop("Thinning period thin must be at least 1 and must not exceed the length of chain after burn-in period (m - burn).");
-
-  S4 results("Mlr"); // Create object slda of class Mlr
-
-  const uint32_t D = x.n_rows;
-  const uint16_t pp1 = x.n_cols;
-  const uint32_t chain_outlength = (m - burn) / thin; // Truncates remainder
-
-  arma::mat etam(chain_outlength, pp1);
-  NumericVector sigma2m(chain_outlength);
-  NumericVector loglike(chain_outlength); // Store log-likelihood (up to an additive constant)
-  NumericVector logpost(chain_outlength); // Store log-posterior (up to an additive constant)
-  arma::mat l_pred(chain_outlength, D);
-
-  // Initialize sigma^2
-  double sigma2 = var(y) / 2.0;
-
-  if (verbose) {
-    Rcout << 1 << " eta: " << eta_start.t() << " ~~~~ sigma2: " << sigma2 << "\n";
-  }
-
-  arma::colvec eta(pp1);
-
-  Progress p(m, display_progress);
-  for (uint32_t i = 1; i <= m; i++) {
-
-    // Draw eta
-    try {
-      eta = draw_eta_norm(x, y, sigma2, mu0, sigma0);
-    } catch (std::exception& e) {
-      Rcerr << "Runtime Error: " << e.what() << " while drawing eta vector\n";
-      forward_exception_to_r(e);
-    }
-
-    // Draw sigma2
-    try {
-      sigma2 = draw_sigma2(a0, b0, x, y, eta);
-    } catch (std::exception& e) {
-      Rcerr << "Runtime Error: " << e.what() << " while drawing sigma2\n";
-      forward_exception_to_r(e);
-    }
-
-    if ( (i > burn) && (i % thin == 0) ) {
-      // Likelihood
-      uint32_t temp_pos = (i - burn - 1) / thin;
-      if (temp_pos == chain_outlength) break; // Invalid index
-      loglike(temp_pos) = get_ll_mlr(y, x, eta, sigma2);
-      // Log-posterior
-      logpost(temp_pos) = get_lpost_mlr(loglike(temp_pos), eta, sigma2,
-                                            mu0, sigma0, a0, b0);
-
-      l_pred.row(temp_pos) = post_pred_norm(x, eta, sigma2);
-
-      etam.row(temp_pos) = eta.t();
-      sigma2m(temp_pos) = sigma2;
-    }
-    if (i % 500 == 0) {
-      if (verbose) {
-        Rcout << i << " eta: " << eta.t() << " ~~~~ sigma2: " << sigma2 << "\n";
-      }
-    }
-    if (display_progress) {
-      p.increment();
-    }
-    Rcpp::checkUserInterrupt(); // Check to see if user cancelled sampler
-  }
-
-  // Compute WAIC and p_eff
-  NumericVector waic_and_se = waic_all(chain_outlength, l_pred);
-
-  results.slot("ndocs") = D;
-  results.slot("nchain") = chain_outlength;
-  results.slot("eta") = etam;
-  results.slot("sigma2") = sigma2m;
-  results.slot("mu0") = mu0;
-  results.slot("sigma0") = sigma0;
-  results.slot("a0") = a0;
-  results.slot("b0") = b0;
-  results.slot("eta_start") = eta_start;
-  results.slot("loglike") = loglike;
-  results.slot("logpost") = logpost;
-  results.slot("waic") = waic_and_se(0);
-  results.slot("se_waic") = waic_and_se(1);
-  results.slot("p_eff") = waic_and_se(2);
-  results.slot("lpd") = l_pred;
-
-  return results;
-}
-
-// Collapsed Gibbs sampler for logistic regression
-//
-// @param m The number of iterations to run the Gibbs sampler.
-// @param burn The number of iterations to discard as the burn-in period.
-// @param thin The period of iterations to keep after the burn-in period
-//   (default: 1).
-// @param y A D x 1 vector of binary outcomes (0/1) to be predicted.
-// @param x A D x p matrix of additional predictors (no column of 1s for
-//   intercept).
-// @param mu0 A (p + 1) x 1 mean vector for the prior on the regression coefficients.
-// @param sigma0 A (p + 1) x (p + 1) variance-covariance matrix for the prior
-//   on the regression coefficients.
-// @param eta_start A (p + 1) x 1 vector of starting values for the
-//   regression coefficients.
-// @param proposal_sd The proposal standard deviation for drawing the
-//   regression coefficients, N(0, proposal_sd) (default: 2.38, ..., 2.38).
-// @param verbose Should parameter draws be output during sampling? (default:
-//   \code{FALSE}).
-// @param display_progress Should percent progress of sampler be displayed
-//   (default: \code{FALSE}). Recommended that only one of \code{verbose} and
-//   \code{display_progress} be set to \code{TRUE} at any given time.
-//
-// @return An object of class \code{Logistic}.
-//' @export
-// [[Rcpp::export(.gibbs_logistic_cpp)]]
-S4 gibbs_logistic_cpp(uint32_t m, uint32_t burn, uint32_t thin,
-                      const arma::colvec& y, const arma::mat& x,
-                      const arma::colvec& mu0, const arma::mat& sigma0,
-                      arma::colvec eta_start, arma::vec proposal_sd,
-                      bool verbose = false, bool display_progress = false) {
-
-  if (m <= burn) stop("Length of chain m not greater than burn-in period.");
-  if ( (thin == 0) || (thin > (m - burn)) )
-    stop("Thinning period thin must be at least 1 and must not exceed the length of chain after burn-in period (m - burn).");
-
-  S4 results("Logistic"); // Create object slda of class Logistic
-
-  const uint32_t D = y.size();
-  const uint16_t pp1 = x.n_cols;
-  const uint32_t chain_outlength = (m - burn) / thin; // Truncates remainder
-
-  arma::mat etam(chain_outlength, pp1);
-  NumericVector loglike(chain_outlength); // Store log-likelihood (up to an additive constant)
-  NumericVector logpost(chain_outlength); // Store log-posterior (up to an additive constant)
-  arma::mat l_pred(chain_outlength, D);
-
-  arma::vec attempt = arma::vec(pp1, arma::fill::zeros);
-  arma::vec accept  = arma::vec(pp1, arma::fill::zeros);
-  arma::vec acc_rate(pp1);
-
-  arma::colvec eta(pp1);
-
-  Progress prog(m, display_progress);
-  for (uint32_t i = 1; i <= m; i++) {
-
-    // Draw eta
-    try {
-      eta = draw_eta_logit(x, y, eta,
-                           mu0, sigma0, proposal_sd,
-                           attempt, accept);
-    } catch (std::exception& e) {
-      Rcerr << "Runtime Error: " << e.what() <<
-        " while drawing eta vector\n";
-      forward_exception_to_r(e);
-    }
-
-    // Update acceptance rates and tune proposal standard deviations
-    acc_rate = accept / attempt;
-    for (uint16_t j = 0; j < pp1; j++) {
-      if (i < (static_cast<float>(burn) / 5.0) &&
-          attempt(j) >= 50 &&
-          (i % 50 == 0)) {
-        if (acc_rate(j) < 0.159) {
-          // too low acceptance, decrease jumping width
-          proposal_sd(j) = proposal_sd(j) * 0.8;
-        }
-        if (acc_rate(j) > 0.309) {
-          // too high acceptance, increase jumping width
-          proposal_sd(j) = proposal_sd(j) * 1.2;
-        }
-        accept(j) = attempt(j) = 0;
-      }
-    }
-
-    if ( (i > burn) && (i % thin == 0) ) {
-      uint32_t temp_pos = (i - burn - 1) / thin;
-      if (temp_pos == chain_outlength) break; // Invalid index
-      // Likelihood
-      loglike(temp_pos) = get_ll_logit(y, x, eta);
-      // Log-posterior
-      logpost(temp_pos) = get_lpost_eta(loglike(temp_pos), eta, mu0, sigma0);
-      l_pred.row(temp_pos) = post_pred_logit(x, eta);
-      etam.row(temp_pos) = eta.t();
-    }
-    if (i % 500 == 0) {
-      if (verbose) {
-        Rcout << i << " eta: " << eta.t() << "\n" <<
-          "AR: " << acc_rate.t() << "\n" <<
-          "Proposal SD: " << proposal_sd.t() << "\n";
-      }
-    }
-    if (display_progress) {
-      prog.increment();
-    }
-    Rcpp::checkUserInterrupt(); // Check to see if user cancelled sampler
-  }
-
-  // Compute WAIC and p_eff
-  NumericVector waic_and_se = waic_all(chain_outlength, l_pred);
-
-  results.slot("ndocs") = D;
-  results.slot("nchain") = chain_outlength;
-  results.slot("eta") = etam;
-  results.slot("mu0") = mu0;
-  results.slot("sigma0") = sigma0;
-  results.slot("eta_start") = eta_start;
-  results.slot("proposal_sd") = proposal_sd;
-  results.slot("loglike") = loglike;
-  results.slot("logpost") = logpost;
-  results.slot("waic") = waic_and_se(0);
-  results.slot("se_waic") = waic_and_se(1);
-  results.slot("p_eff") = waic_and_se(2);
-  results.slot("lpd") = l_pred;
-
-  return results;
-}
-
-// Collapsed Gibbs sampler for the sLDA-X model
-//
-// @param m The number of iterations to run the Gibbs sampler.
-// @param burn The number of iterations to discard as the burn-in period.
-// @param thin The period of iterations to keep after the burn-in period
-//   (default: 1).
-// @param y A D x 1 vector of binary outcomes (0/1) to be predicted.
-// @param x A D x p matrix of additional predictors (no column of 1s for
-//   intercept).
-// @param docs A D x max(\eqn{N_d}) matrix of word indices for all documents.
-// @param V The number of unique terms in the vocabulary.
-// @param K The number of topics.
-// @param model An integer denoting the type of model to fit.
-// @param mu0 A K x 1 mean vector for the prior on the regression coefficients.
-// @param sigma0 A (K + p + 1) x (K + p + 1) variance-covariance matrix for the
-//   prior on the regression coefficients. The first p + 1 columns/rows
-//   correspond to predictors in X, while the last K columns/rows correspond to
-//   the K topic means.
-// @param a0 The shape parameter for the prior on sigma2.
-// @param b0 The scale parameter for the prior on sigma2.
-// @param eta_start A (K + p) x 1 vector of starting values for the
-//   regression coefficients. The first p elements correspond to predictors
-//   in X, while the last K elements correspond to the K topic means.
-// @param constrain_eta A logical (default = \code{TRUE}): If \code{TRUE}, the
-//   regression coefficients will be constrained so that they are in descending
-//   order; if \code{FALSE}, no constraints will be applied.
-// @param alpha_ The hyper-parameter for the prior on the topic proportions
-//   (default: 0.1).
-// @param gamma_ The hyper-parameter for the prior on the topic-specific
-//   vocabulary probabilities (default: 1.01).
-// @param proposal_sd The proposal standard deviation for drawing the
-//   regression coefficients, N(0, proposal_sd) (default: 0.2).
-// @param interaction_xcol The column number of the design matrix for the
-// additional predictors for which an interaction with the \eqn{K} topics is
-// desired (default: \eqn{-1L}, no interaction). Currently only supports a
-// single continuous predictor or a two-category categorical predictor
-// represented as a single dummy-coded column.
-// @param verbose Should parameter draws be output during sampling? (default:
-//   \code{FALSE}).
-// @param display_progress Should percent progress of sampler be displayed
-//   (default: \code{FALSE}). Recommended that only one of \code{verbose} and
-//   \code{display_progress} be set to \code{TRUE} at any given time.
-//' @export
-// [[Rcpp::export(.gibbs_sldax_cpp)]]
-S4 gibbs_sldax_cpp(const arma::umat& docs, uint32_t V,
-                   uint32_t m, uint32_t burn, uint32_t thin,
-                   uint16_t K, uint8_t model,
-                   const arma::colvec& y,
-                   const arma::mat& x,
-                   const arma::colvec& mu0,
-                   const arma::mat& sigma0,
-                   float a0, float b0,
-                   arma::colvec eta_start,
-                   arma::vec proposal_sd,
-                   int interaction_xcol = -1,
-                   float alpha_ = 0.1, float gamma_ = 1.01,
-                   bool constrain_eta = true,
-                   bool verbose = false, bool display_progress = false) {
-
-  if (m <= burn) stop("Length of chain m not greater than burn-in period.");
-  if ( (thin == 0) || (thin > (m - burn)) )
-    stop("Thinning period thin must be at least 1 and must not exceed the length of chain after burn-in period (m - burn).");
-
-  const uint8_t lda = 1;
-  const uint8_t slda = 2;
-  const uint8_t sldax = 3;
-  const uint8_t slda_logit = 4;
-  const uint8_t sldax_logit = 5;
-
-  S4 results("Sldax");
-
-  const uint32_t chain_outlength = (m - burn) / thin; // Truncates remainder
-  NumericVector sigma2m(chain_outlength);
-  double sigma2 = var(y) / 2.0;
-
-  const uint32_t D = docs.n_rows;
-  arma::colvec N(D);
-  const IntegerVector topics_index = seq_len(K);
-  const IntegerVector docs_index   = seq_len(D) - 1;
-  for (uint32_t d : docs_index) N(d) = sum(docs.row(d) > 0);
-  const uint32_t maxNd = max(N);
-  arma::mat theta(D, K);
-  arma::mat beta(K, V);
-  // Topic draw counts
-  arma::mat ndk(D, K);
-  // Counts of topic-word co-occurences in corpus (K x V)
-  arma::mat nkm(K, V);
-  // Topic draws for all words and docs
-  arma::vec nk(K);
-  arma::umat zdocs = arma::umat(D, maxNd, arma::fill::zeros);
-  arma::mat zbar(D, K);
-
-  // D x max(N_d) x m array to store topic draws
-  arma::ucube topicsm = arma::ucube(D, maxNd, chain_outlength, arma::fill::zeros);
-  NumericVector loglike(chain_outlength); // Store log-likelihood (up to an additive constant)
-  NumericVector logpost(chain_outlength); // Store log-posterior (up to an additive constant)
-
-  for (uint32_t d : docs_index) {
-    // Randomly initialize topic draws from Uniform(1, K)
-    zdocs.row(d).subvec(0, N(d) - 1) = arma::randi<arma::urowvec>( N(d), arma::distr_param(1, K) );
-    for (uint16_t k = 0; k < K; k++) {
-      // Count topic draws in each document
-      ndk(d, k) = sum(zdocs.row(d) == (k + 1));
-    }
-  }
-
-  if (model != lda) {
-    // Get zbar matrix
-    zbar = ndk.each_col() / N;
-  }
-
-  try {
-    nkm = count_topic_word(K, V, zdocs, docs);
-  } catch(std::exception& e) {
-    Rcerr << "Runtime error: " << e.what() <<
-      " while computing topic-word co-occurrences\n";
-    forward_exception_to_r(e);
-  }
-
-  nk = arma::sum(ndk, 0).t(); // Column sums
-
-  // Estimate theta
-  for (uint32_t d : docs_index) {
-    try {
-      theta.row(d) = est_thetad(ndk.row(d), alpha_);
-    } catch(std::exception& e) {
-      Rcerr << "Runtime Error: " << e.what() <<
-        " when estimating theta vector for document " << d << "\n";
-      forward_exception_to_r(e);
-    }
-  }
-
-  // Estimate beta
-  for (uint16_t k = 0; k < K; k++) {
-    try {
-      beta.row(k) = est_betak(nkm.row(k), gamma_);
-    } catch(std::exception& e) {
-      Rcerr << "Runtime Error: " << e.what() << " estimating row " << k <<
-        "of beta matrix\n";
-      forward_exception_to_r(e);
-    }
-  }
-
-  // For supervised models, set number of columns in design matrix
-  uint16_t p = 0; // Num. cols. involving fixed predictors
-  uint16_t q = 0; // Total num. cols. in design matrix
-  if (model == sldax || model == sldax_logit) {
-    if (interaction_xcol > 0) {
-      p = x.n_cols + K - 1; // Need K - 1 interaction columns
-    } else {
-      p = x.n_cols;
-    }
-    q = p + K; // num. covariates + K + (K - 1) for interaction of one covariate w/ topics
-  } else if (model == slda || model == slda_logit) {
-    p = K; // Num. cols. involving fixed predictors
-    q = p; // Total num. cols. in design matrix
-  }
-  arma::mat etam(chain_outlength, q);
-  arma::colvec eta(q);
-  arma::colvec etac(q);
-
-  // For supervised models sLDA/sLDAX, set up regression coefficients
-  if (constrain_eta) {
-    if (model == sldax || model == sldax_logit) {
-      // Constrain starting values of eta to be in descending order
-      if (interaction_xcol > 0) {
-        // Only sort topic "linear effects", not interactions
-        std::partial_sort(eta_start.begin() + p - K + 1,
-                          eta_start.end() - K + 1,
-                          eta_start.end(), std::greater<float>());
-      } else {
-        std::sort(eta_start.begin() + p,
-                  eta_start.end(),
-                  std::greater<float>());
-      }
-    } else if (model == slda || model == slda_logit) {
-      std::sort(eta_start.begin(), eta_start.end(), std::greater<float>());
-    }
-  }
-
-  if (verbose && model != lda) Rcout << 1 << " eta: " << eta_start.t() << "\n";
-  eta = eta_start;
-
-  arma::vec attempt = arma::vec(q, arma::fill::zeros);
-  arma::vec accept  = arma::vec(q, arma::fill::zeros);
-  arma::vec acc_rate(q);
-
-  arma::mat r = zbar;
-  // Predictive posterior likelihood of y
-  arma::mat l_pred(chain_outlength, D);
-  arma::mat xz_int(D, K - 1); // Need K - 1 interaction columns
-  if (model == sldax || model == sldax_logit) {
-    r = join_rows(x, zbar); // Full predictor matrix `r`
-    if (interaction_xcol > 0) {
-      xz_int = zbar.cols(0, K - 2).each_col() % x.col(interaction_xcol - 1);
-      r = join_rows(r, xz_int);
-    }
-  }
-
-  Progress prog(m, display_progress);
-  for (uint32_t i = 1; i <= m; i++) {
-
-    // Draw z
-    for (uint32_t d : docs_index) {
-      for (uint32_t n = 0; n < N(d); n++) {
-        uint32_t word = docs(d, n);
-        uint16_t topic = zdocs(d, n);
-        // Draw topic z_{dn} excluding current z_{dn} value
-        update_zcounts(d, word, topic, d, ndk, nk, nkm);
-        try {
-          if (model == lda) {
-            topic = draw_zdn_lda(V, ndk.row(d).t(), nkm.col(word - 1), nk,
-                                 alpha_, gamma_);
-          } else if (model == slda || model == sldax) {
-            topic = draw_zdn_slda_norm(y(d), r.row(d), eta, sigma2, V,
-                                       ndk.row(d).t(), nkm.col(word - 1), nk,
-                                       alpha_, gamma_);
-          } else if (model == slda_logit || model == sldax_logit) {
-            topic = draw_zdn_slda_logit(y(d), r.row(d), eta, V,
-                                        ndk.row(d).t(), nkm.col(word - 1), nk,
-                                        alpha_, gamma_);
-          } else {
-            Rcerr << "Invalid model specified. Exiting.\n";
-          }
-        } catch(std::exception& e) {
-          Rcerr << "Runtime Error: " << e.what() <<
-            " occurred while drawing topic for word " << n << " in document "
-            << d << "\n";
-          forward_exception_to_r(e);
-        }
-
-        zdocs(d, n) = topic;
-        // Update topic count in doc d
-        ndk(d, topic - 1)++;
-        // Update topic count in corpus
-        nk(topic - 1)++;
-        // Update topic-word counts in corpus
-        nkm(topic - 1, word - 1)++;
-      }
-    }
-
-    // Estimate theta for doc d
-    for (uint32_t d: docs_index) {
-      try {
-        theta.row(d) = est_thetad(ndk.row(d), alpha_);
-      } catch(std::exception& e) {
-        Rcerr << "Runtime Error: " << e.what() <<
-          " when estimating theta vector for document " << d << "\n";
-        forward_exception_to_r(e);
-      }
-    }
-
-    if (model != lda) {
-      // Get zbar matrix
-      zbar = ndk.each_col() / N;
-    }
-
-    if (model == sldax || model == sldax_logit) {
-      r = join_rows(x, zbar); // Full predictor matrix `r`
-      if (interaction_xcol > 0) {
-        xz_int = zbar.cols(0, K - 2).each_col() % x.col(interaction_xcol - 1);
-        r = join_rows(r, xz_int);
-      }
-    } else {
-      r = zbar;
-    }
-
-    // Estimate beta
-    for (uint16_t k = 0; k < K; k++) {
-      try {
-        beta.row(k) = est_betak(nkm.row(k), gamma_);
-      } catch(std::exception& e) {
-        Rcerr << "Runtime Error: " << e.what() << " estimating row " << k <<
-          "of beta matrix\n";
-        forward_exception_to_r(e);
-      }
-    }
-
-    // Draw eta
-    if (model != lda) {
-      bool eta_order = false;
-      uint16_t iter = 0;
-      uint16_t max_iter = 1; // Only draw once
-      if (constrain_eta) {
-        max_iter = 10000;   // Max repeated draws if trying to constrain eta
-      }
-      while (!eta_order & (iter < max_iter)) { // Runs at least once
-        iter++;
-        if (model == slda || model == sldax) {
-          try {
-            etac = draw_eta_norm(r, y, sigma2, mu0, sigma0);
-          } catch (std::exception& e) {
-            Rcerr << "Runtime Error: " << e.what() <<
-              " while drawing eta vector\n";
-            forward_exception_to_r(e);
-          }
-        } else if (model == slda_logit || model == sldax_logit) {
-          try {
-            etac = draw_eta_logit(r, y, eta, mu0, sigma0,
-                                  proposal_sd, attempt, accept);
-          } catch (std::exception& e) {
-            Rcerr << "Runtime Error: " << e.what() <<
-              " while drawing eta vector\n";
-            forward_exception_to_r(e);
-          }
-        }
-        if (constrain_eta) {
-          if (interaction_xcol > 0) {
-            for (uint16_t k = p - K + 2; k < p + 1; k++) {
-              // Force eta components to be in descending order (first is largest) to resolve label switching of topics
-              eta_order = (etac(k - 1) > etac(k)) || (abs(etac(k - 1) - etac(k)) < .000001);
-              if (!eta_order) break;
-            }
-          } else {
-            for (uint16_t k = p + 1; k < q; k++) {
-              // Force eta components to be in descending order (first is largest)
-              //   to resolve label switching of topics
-              eta_order = (etac(k - 1) > etac(k)) || (abs(etac(k - 1) - etac(k)) < .000001);
-              if (!eta_order) break;
-            }
-          }
-        }
-      }
-      if (iter >= max_iter - 1 & max_iter > 1) Rcout << "Reached max_iter while drawing eta\n";
-      eta = etac; // New draw
-
-      if (model == slda || model == sldax) {
-        // Draw sigma2
-        try {
-          sigma2 = draw_sigma2(a0, b0, r, y, eta);
-        } catch (std::exception& e) {
-          Rcerr << "Runtime Error: " << e.what() << " while drawing sigma2\n";
-          forward_exception_to_r(e);
-        }
-      }
-    }
-
-    if ( (i > burn) && (i % thin == 0) ) {
-      if (i == burn + 1 && burn > 0) Rcout << "Finished burn-in period\n";
-
-      uint32_t temp_pos = (i - burn - 1) / thin;
-      if (temp_pos == chain_outlength) break; // Invalid index
-
-      topicsm.slice(temp_pos) = zdocs;
-      // Likelihood
-      switch(model) {
-        case lda:
-          loglike(temp_pos) = get_ll_lda(zdocs, docs, theta, beta,
-                                             docs_index, N);
-          logpost(temp_pos) = get_lpost_lda(
-            loglike(temp_pos), theta, beta, gamma_, alpha_, V, docs_index);
-          break;
-        case slda: // Fall through
-        case sldax:
-          loglike(temp_pos) = get_ll_slda_norm(
-            y, r, eta, sigma2, zdocs, docs, theta, beta, docs_index, N);
-          logpost(temp_pos) = get_lpost_slda_norm(
-            loglike(temp_pos), eta, sigma2, theta, beta, mu0, sigma0, gamma_,
-            alpha_, a0, b0, V, docs_index);
-          l_pred.row(temp_pos) = post_pred_norm(r, eta, sigma2);
-          sigma2m(temp_pos) = sigma2;
-          break;
-        case slda_logit: // Fall through
-        case sldax_logit:
-          loglike(temp_pos) = get_ll_slda_logit(
-            y, r, eta, zdocs, docs, theta, beta, docs_index, N);
-          logpost(temp_pos) = get_lpost_slda_logit(loglike(temp_pos),
-            eta, theta, beta, mu0, sigma0, gamma_, alpha_, V, docs_index);
-          l_pred.row(temp_pos) = post_pred_logit(r, eta);
-          break;
-        default: Rcerr << "Invalid model specified. Exiting.\n";
-      }
-      if (model != lda) {
-        etam.row(temp_pos) = eta.t();
-      }
-    }
-
-    if (model == slda_logit || model == sldax_logit) {
-      acc_rate = accept / attempt;
-      for (uint16_t j = 0; j < q; j++) {
-        if (i < (static_cast<float>(burn) / 5.0) &&
-            attempt(j) >= 50 &&
-            (i % 50 == 0)) {
-          if (acc_rate(j) < 0.159) {
-            // too low acceptance, decrease jumping width
-            proposal_sd(j) = proposal_sd(j) * 0.8;
-          }
-          if (acc_rate(j) > 0.309) {
-            // too high acceptance, increase jumping width
-            proposal_sd(j) = proposal_sd(j) * 1.2;
-          }
-          accept(j) = attempt(j) = 0;
-        }
-      }
-      if (i == burn / 5.0 && burn > 0) Rcout << "Finished tuning proposal distributions\n";
-    }
-
-    if (i % 500 == 0 && verbose) {
-      switch(model) {
-        case slda: // Fall through
-        case sldax:
-          Rcout << i << " eta: " << eta.t() <<
-            " ~~~~ sigma2: " << sigma2 <<
-            " ~~~~ zbar d1: " << zbar.row(0) << "\n";
-          break;
-        case slda_logit: // Fall through
-        case sldax_logit:
-          Rcout << i << " eta: " << eta.t() <<
-            " ~~~~ zbar d1: " << zbar.row(0) <<
-            " ~~~~ AR: " << acc_rate.t() << "\n" <<
-            " ~~~~ Proposal SD: " << proposal_sd.t() << "\n";
-          break;
-        default:
-          break;
-      }
-    }
-    if (display_progress) {
-      prog.increment();
-    }
-    Rcpp::checkUserInterrupt(); // Check to see if user cancelled sampler
-  }
-
-  // Compute WAIC and p_eff
-  NumericVector waic_and_se = waic_all(chain_outlength, l_pred);
-
-  results.slot("ntopics") = K;
-  results.slot("ndocs")   = D;
-  results.slot("nvocab")  = V;
-  results.slot("nchain")  = chain_outlength;
-  results.slot("topics")  = topicsm;
-  results.slot("alpha")   = alpha_;
-  results.slot("gamma")   = gamma_;
-  results.slot("loglike") = loglike;
-  results.slot("logpost") = logpost;
-
-  if (model != lda) {
-    results.slot("eta")       = etam;
-    results.slot("mu0")       = mu0;
-    results.slot("sigma0")    = sigma0;
-    results.slot("eta_start") = eta_start;
-    results.slot("waic")      = waic_and_se(0);
-    results.slot("se_waic")   = waic_and_se(1);
-    results.slot("p_eff")     = waic_and_se(2);
-    results.slot("lpd")       = l_pred;
-
-    if (model == slda || model == sldax) {
-      results.slot("sigma2") = sigma2m;
-      results.slot("a0")     = a0;
-      results.slot("b0")     = b0;
-    }
-
-    if (model == slda_logit || model == sldax_logit)
-      results.slot("proposal_sd") = proposal_sd;
-  }
-
-  return results;
-}
